EM-2852: Push functional setup changes
