+++ bin/start-local-environment.sh
@@ -39,6 +39,10 @@ done
 # Set up IDAM client with services and roles
 echo "Setting up IDAM client..."
 (./bin/idam-client-setup.sh ${IDAM_URI} services ${token} '{"description": "em", "label": "em", "oauth2ClientId": "webshow", "oauth2ClientSecret": "AAAAAAAAAAAAAAAA", "oauth2RedirectUris": ["http://localhost:8080/oauth2redirect"], "selfRegistrationAllowed": true}')
+(./bin/idam-client-setup.sh ${IDAM_URI} services ${token} '{"description": "ccd gateway", "label": "ccd gateway", "oauth2ClientId": "ccd_gateway", "oauth2ClientSecret": "AAAAAAAAAAAAAAAA", "oauth2RedirectUris": ["http://localhost:3451/oauth2redirect"], "selfRegistrationAllowed": true}')
+(./bin/idam-client-setup-roles.sh ${IDAM_URI} ${token} caseworker)
+(./bin/idam-client-setup-roles.sh ${IDAM_URI} ${token} caseworker-publiclaw)
+(./bin/idam-client-setup-roles.sh ${IDAM_URI} ${token} ccd-import)
 
 # Start all other images
 echo "Starting dependencies..."
@@ -46,10 +50,8 @@ docker-compose ${COMPOSE_FILE} build
 docker-compose ${COMPOSE_FILE} up -d shared-database\
                                      service-auth-provider-api \
                                      smtp-server \
-                                     rpa-native-pdf-annotator-db \
-                                     annotation-api-db \
-                                     dm-store-db \
                                      dm-store \
+                                     rpa-native-pdf-annotator-db \
                                      ccd-user-profile-api \
                                      ccd-definition-store-api \
                                      ccd-data-store-api \
