+++ docker-compose-dependencies.yml
@@ -20,7 +20,11 @@ services:
       - jwtKey=wThK0f0/lh3FlxFcL4xUWDMI5C1J9KyQBgXV4wseh1e5J1uYJIjvTvArHxQDrYoHJ23xFxjHkOnvNbR5dXRoxA==
       - microserviceKeys_em_gw=AAAAAAAAAAAAAAAA
       - microserviceKeys_em_npa_app=AAAAAAAAAAAAAAAA
-#      logging env vars
+      - microserviceKeys_ccd_data=AAAAAAAAAAAAAAAA
+      - microserviceKeys_ccd_gw=AAAAAAAAAAAAAAAA
+      - microserviceKeys_ccd_definition=AAAAAAAAAAAAAAAA
+      - microserviceKeys_ccd_admin=AAAAAAAAAAAAAAAA
+      #      logging env vars
       - JSON_CONSOLE_PRETTY_PRINT=false
       - REFORM_SERVICE_TYPE=java
       - REFORM_SERVICE_NAME=service-auth-provider-api
@@ -39,28 +43,17 @@ services:
     image: hmctspublic.azurecr.io/em/anno:latest
     container_name: npa-anno-api
     environment:
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://annotation-api-db:5432/anno
-      - SPRING_DATASOURCE_USERNAME=anno
-      - SPRING_DATASOURCE_PASSWORD=anno
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://shared-database:5432/anno
+      - SPRING_DATASOURCE_USERNAME=shared
+      - SPRING_DATASOURCE_PASSWORD=shared
       - S2S_BASE_URI=http://service-auth-provider-api:8489
       - IDAM_API_BASE_URI=http://idam-api:5000
       - OPEN_ID_API_BASE_URI=http://idam-api:5000/o
       - ENABLE_LIQUIBASE=true
-    links:
-    - annotation-api-db
     depends_on:
-    - annotation-api-db
+    - shared-database
     ports:
     - 4623:8080
-
-  annotation-api-db:
-    container_name: npa-anno-db
-    image: postgres
-    restart: always
-    environment:
-      - POSTGRES_DB=anno
-      - POSTGRES_USER=anno
-      - POSTGRES_PASSWORD=anno
 
   idam-api:
     image: hmctspublic.azurecr.io/idam/api:stable
@@ -140,10 +133,9 @@ services:
       - IDAM_SUPERUSER_EMAIL
       - IDAM_SECRET_KEY=AAAAAAAAAAAAAAAA
       - GOV_NOTIFY_API_KEY
-      - ENABLE_DB_MIGRATE=true
-      - SPRING_DATASOURCE_URL=jdbc:postgresql://dm-store-db:5432/evidence
-      - SPRING_DATASOURCE_USERNAME=evidence
-      - SPRING_DATASOURCE_PASSWORD=evidence
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://shared-database:5432/evidence
+      - SPRING_DATASOURCE_USERNAME=shared
+      - SPRING_DATASOURCE_PASSWORD=shared
       - SPRING_PROFILES_ACTIVE=dev
       - ENABLE_AZURE_STORAGE_CONTAINER=true
       - ENABLE_POSTGRES_BLOB_STORAGE=false
@@ -170,28 +162,14 @@ services:
       - PACKAGES_NAME=dm-store
       - PACKAGES_VERSION=unkown
       #      debug mode
-      - JAVA_OPTS=-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=n,address=5005
-    links:
-      - dm-store-db
-      - service-auth-provider-api
-      - make-container-call
+      #- JAVA_TOOL_OPTIONS=-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=n,address=5005
+      - JAVA_TOOL_OPTIONS=-Xms128m -Xmx512m -XX:MaxPermSize=512m
     depends_on:
-      - dm-store-db
+      - shared-database
       - service-auth-provider-api
       - make-container-call
     ports:
       - 4603:8080
-
-  dm-store-db:
-    image: postgres
-    container_name: npa-dm-store-db
-    restart: always
-    environment:
-      - POSTGRES_DB=evidence
-      - POSTGRES_USER=evidence
-      - POSTGRES_PASSWORD=evidence
-    ports:
-      - 5435:5432
 
   ccd-user-profile-api:
     image: hmctspublic.azurecr.io/ccd/user-profile-api:latest
@@ -202,7 +180,7 @@ services:
       USER_PROFILE_DB_USERNAME: shared
       USER_PROFILE_DB_PASSWORD: shared
       USER_PROFILE_DB_USE_SSL: "false"
-      USER_PROFILE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_definition,ccd_admin
+      USER_PROFILE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_definition,ccd_admin,em_npa_app
       IDAM_S2S_URL: http://service-auth-provider-api:8489
       REFORM_SERVICE_NAME: ccd-user-profile-api
       REFORM_TEAM: ccd
@@ -225,7 +203,7 @@ services:
       DEFINITION_STORE_DB_PASSWORD: shared
       DEFINITION_STORE_DB_USE_SSL: "false"
       DEFINITION_STORE_IDAM_KEY: AAAAAAAAAAAAAAAA
-      DEFINITION_STORE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_gw,ccd_admin
+      DEFINITION_STORE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_gw,ccd_admin,em_npa_app
       USER_PROFILE_HOST: http://ccd-user-profile-api:4453
       IDAM_USER_URL: http://idam-api:5000
       IDAM_S2S_URL: http://service-auth-provider-api:8489
