+++ src/aat/java/uk/gov/hmcts/reform/em/npa/functional/BaseTest.java
@@ -1,0 +1,22 @@
+package uk.gov.hmcts.reform.em.npa.functional;
+
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
+import uk.gov.hmcts.reform.em.EmTestConfig;
+import uk.gov.hmcts.reform.em.npa.testutil.TestUtil;
+
+@SpringBootTest(classes = {TestUtil.class, EmTestConfig.class})
+@PropertySource(value = "classpath:application.yml")
+@RunWith(SpringRunner.class)
+public abstract class BaseTest {
+
+    @Autowired
+    protected TestUtil testUtil;
+
+    @Autowired
+    private CoreCaseDataApi coreCaseDataApi;
+}
