+++ src/aat/resources/application.yml
@@ -31,6 +31,8 @@ s2s:
     url: ${S2S_BASE_URI:http://localhost:4502}
     secret: ${FUNCTIONAL_TEST_CLIENT_S2S_TOKEN:AAAAAAAAAAAAAAAA}
     serviceName: ${S2S_SERVICE_NAME:em_gw}
+    ccdGwSecret: ${FUNCTIONAL_TEST_CCD_GW_SERVICE_SECRET:AAAAAAAAAAAAAAAA}
+    ccdGwServiceName: ${CCD_GW_SERVICE_NAME:ccd_gw}
 
 ccd-def:
   api:
@@ -38,6 +40,10 @@ ccd-def:
 
 document_management:
   url: ${DM_STORE_APP_URL:http://localhost:4603}
+
+core_case_data:
+  api:
+    url: ${CCD_DATA_API:http://localhost:4452}
 
 ccd:
   data:
